requirements.txt
Flask==2.3.2
ccxt==4.4.51
numpy==2.2.1
python-dotenv==1.0.1
gunicorn==23.1.0